pyproject.toml
[tool.poetry]
name = "gitmess"
version = "0.1.0"
description = ""
authors = ["0xjesus <pdlgmcn@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.300"
gitpython = "^3.1.37"
openai = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gm = 'gitmess.gitmess:fetch_git_changes'

